Move cube/fit.py to datasets/fit.py
