+++ gammapy/cube/__init__.py
@@ -3,7 +3,6 @@
 from .background import *
 from .edisp_map import *
 from .exposure import *
-from .fit import *
 from .fov import *
 from .make import *
 from .psf_kernel import *
