+++ gammapy/datasets/core.py
@@ -6,7 +6,7 @@ import numpy as np
 from gammapy.utils.scripts import make_name, make_path, read_yaml, write_yaml
 from gammapy.utils.table import table_from_row_data
 from gammapy.maps import WcsNDMap
-from gammapy.modeling.parameter import Parameters
+
 
 __all__ = ["Dataset", "Datasets"]
 
@@ -123,6 +123,9 @@ class Datasets(collections.abc.MutableSequence):
         Duplicate parameter objects have been removed.
         The order of the unique parameters remains.
         """
+        #TODO: remove this delayed import
+        from gammapy.modeling.parameter import Parameters
+
         parameters = Parameters.from_stack([_.models.parameters for _ in self])
         return parameters.unique_parameters
