+++ gammapy/datasets/tests/test_fit.py
@@ -7,11 +7,10 @@ from astropy.coordinates import SkyCoord
 from regions import CircleSkyRegion
 from gammapy.cube import (
     EDispMap,
-    MapDataset,
-    MapDatasetOnOff,
     PSFMap,
     make_map_exposure_true_energy,
 )
+from gammapy.datasets import MapDataset, MapDatasetOnOff
 from gammapy.data import GTI
 from gammapy.irf import EffectiveAreaTable2D, EnergyDependentMultiGaussPSF
 from gammapy.maps import Map, MapAxis, WcsGeom, WcsNDMap
