+++ gammapy/detect/tests/test_asmooth.py
@@ -3,7 +3,7 @@ import pytest
 from numpy.testing import assert_allclose
 import astropy.units as u
 from astropy.convolution import Tophat2DKernel
-from gammapy.cube import MapDatasetOnOff
+from gammapy.datasets import MapDatasetOnOff
 from gammapy.detect import ASmoothMapEstimator
 from gammapy.maps import Map, WcsNDMap
 from gammapy.datasets import Datasets
