+++ gammapy/detect/tests/test_lima.py
@@ -5,7 +5,7 @@ import astropy.units as u
 from astropy.convolution import Tophat2DKernel
 from gammapy.detect import LiMaMapEstimator
 from gammapy.maps import Map, MapAxis, WcsGeom
-from gammapy.cube import MapDataset, MapDatasetOnOff
+from gammapy.datasets import MapDataset, MapDatasetOnOff
 from gammapy.utils.testing import requires_data
